requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0  # 用于天气图表
altair>=5.0.0  # 用于天气图表

crewai==0.28.8
langchain-community==0.0.20
streamlit==1.29.0
python-dotenv==1.0.0